requirements.txt
autopep8==2.0.2
django==4.2.2
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
python-decouple==3.8
